server/requirements.txt
beautifulsoup4==4.12.3
fastapi==0.115.7
httpx==0.28.1
pymongo==4.10.1
python-dotenv==1.0.1
rich==13.9.4
uvicorn==0.34.0
opencv-python==4.9.0.80
